refactor: 减少绑定通知信息的参数层级
并将其移动到合理的文件夹内
